refactor: rename internal modules with a `_` prefix
