+++ copier/__main__.py
@@ -1,6 +1,6 @@
 """Copier CLI entrypoint."""
 
-from .cli import CopierApp
+from ._cli import CopierApp
 
 if __name__ == "__main__":
     CopierApp.run()
