+++ copier/errors.py
@@ -9,12 +9,12 @@ from pathlib import Path
 from subprocess import CompletedProcess
 from typing import TYPE_CHECKING
 
-from .tools import printf_exception
-from .types import PathSeq
+from ._tools import printf_exception
+from ._types import PathSeq
 
 if TYPE_CHECKING:  # always false
-    from .template import Template
-    from .user_data import AnswersMap, Question
+    from ._template import Template
+    from ._user_data import AnswersMap, Question
 
 if sys.version_info < (3, 11):
     from typing_extensions import Self
