+++ docs/reference/jinja_ext.md
@@ -1,1 +1,1 @@
-::: copier.jinja_ext
+::: copier._jinja_ext
