+++ docs/reference/subproject.md
@@ -1,1 +1,1 @@
-::: copier.subproject
+::: copier._subproject
