+++ docs/reference/template.md
@@ -1,1 +1,1 @@
-::: copier.template
+::: copier._template
