+++ docs/reference/tools.md
@@ -1,1 +1,1 @@
-::: copier.tools
+::: copier._tools
