+++ docs/reference/types.md
@@ -1,1 +1,1 @@
-::: copier.types
+::: copier._types
