+++ docs/reference/user_data.md
@@ -1,1 +1,1 @@
-::: copier.user_data
+::: copier._user_data
