+++ docs/reference/vcs.md
@@ -1,1 +1,1 @@
-::: copier.vcs
+::: copier._vcs
