+++ tests/test_legacy_migration.py
@@ -7,8 +7,8 @@ import pytest
 from plumbum import local
 
 from copier import run_copy, run_update
+from copier._user_data import load_answersfile_data
 from copier.errors import UserMessageError
-from copier.user_data import load_answersfile_data
 
 from .helpers import BRACKET_ENVOPS_JSON, PROJECT_TEMPLATE, build_file_tree, git
