+++ tests/test_subdirectory.py
@@ -6,7 +6,7 @@ import pytest
 from plumbum import local
 
 import copier
-from copier.user_data import load_answersfile_data
+from copier._user_data import load_answersfile_data
 
 from .helpers import BRACKET_ENVOPS_JSON, SUFFIX_TMPL, build_file_tree, git, git_init
