+++ tests/test_templated_prompt.py
@@ -11,9 +11,9 @@ from pexpect.popen_spawn import PopenSpawn
 from plumbum import local
 
 from copier import Worker
+from copier._types import AnyByStrDict
+from copier._user_data import load_answersfile_data
 from copier.errors import InvalidTypeError
-from copier.types import AnyByStrDict
-from copier.user_data import load_answersfile_data
 
 from .helpers import (
     BRACKET_ENVOPS,
